Kevin's New Stuff!
-In the Draw class new PNG have been loaded for the currency display and
the death display.
-The Enemy class now has code to add 10 coins to the player when an
enemy is killed.
-The Fillers.png has been edited to where it now includes a small
graphic for enemy health bars.
-The code in the Game class for money is no longer used however when I
tired to remove it, it created errors within the scripts and pause
state, so I decided to leave it there.
-GameHandler now has a new state in it, Death.
-The HUD class no longer uses the Draw method for drawing the numbers,
it uses the TextHandler, and it now includes a constructor for a HUD
object which holds each characters health, and money. Also code fore
displaying the player's currency.
-The Player class now has code to switch characters when a character
dies, however, the Animations still need to be worked out for death.
There are also conditionals for preventing movement and animations by a
dead character. There are also some new methods to return the current
character, their attack, defense, HUD object, magic, level, and XP.
-New state in the State Class, Death.
-The Bank class is an entirely new class that adds a single bank object
to the game that any character can deposit into or withdraw from.
(Starting balance 100 coins).
-New Death Class which displays a graphic "You Died" when all characters
have died.
-There is a new PNG with font numbers and a logo for the currency
display.